fix: 자동 업로드 최종 확인을 \"확인\" 단계로 이동, 미리보기 카드 제거, Shorts → Shorts tag (#235)
- UploadSettingsStep(4단계): \"자동 업로드 최종 확인\" Card와 \"미리보기\" Card 제거.
  더 이상 안 쓰는 useAuthStore, useMemo, ReviewItem helper, firstLangName/previewTitle/
  previewDescription/needsAutoUploadReview/privacyLabel/targetChannelLabel 변수 정리.
- TranslationEditStep(5단계, \"확인\"): 자동 업로드 최종 확인 카드 이식. 자동 업로드 ON일 때만
  노출되어 더빙 시작 전 마지막 동의 게이트 역할. 체크 안 하면 \"더빙 시작\" 버튼 비활성.
- ReviewItem 라벨 \"Shorts\" → \"Shorts tag\"로 명확화. 동의 체크박스 문구도 동일하게 정정.

자동 업로드 가드 자체는 유지 — 노출 시점만 4단계에서 5단계로 이동.
